First phase of implementing the concept of squads (groups)
Squads restrict visibility of other users in StepUp to squads (groups) of folks who know each other or who have been invited to join the squad.
Create a MySquads JSP that shows squads users are in, which they own, and which they are invited to.  Implemented changes to the Dashboard so that
it now shows a Squad dashboard and enforces that only activity/posts are visible if users are full members/owner of the squad.
SQL bootstrap file updated with new schema that includes two new tables to support Squads plus a squadID in the Posts table.
Added missing styling for <h2> to css file as well as an emphasis text that could be used for <span> tags in the future.