Implemented owner's invitation functionality to squads.
Added link to invite members from the mySquads JSP and created new JSP
which allows the owner to type in a username of someone they'd like to invite
to a squad and click a button.  Modified StepUpDAO and FrontController to
work with database to add the invited person if it's a valid request.
SquadMembership.java is a POJO that holds invited member name and squad ID for
passign back and forth betw FrontController and DAO.