Add groups branch (#9)
* Added tables Squads and SquadMembers to database

Implemented changes to database structure to enable grouping (squad) functionality. Includes adding two tables and modifying Posts table to have notion of a squad ID the post was made in.

* First phase of implementing the concept of squads (groups)

Squads restrict visibility of other users in StepUp to squads (groups) of folks who know each other or who have been invited to join the squad.
Create a MySquads JSP that shows squads users are in, which they own, and which they are invited to.  Implemented changes to the Dashboard so that
it now shows a Squad dashboard and enforces that only activity/posts are visible if users are full members/owner of the squad.
SQL bootstrap file updated with new schema that includes two new tables to support Squads plus a squadID in the Posts table.
Added missing styling for <h2> to css file as well as an emphasis text that could be used for <span> tags in the future.

* Implemented owner's invitation functionality to squads.

Added link to invite members from the mySquads JSP and created new JSP
which allows the owner to type in a username of someone they'd like to invite
to a squad and click a button.  Modified StepUpDAO and FrontController to
work with database to add the invited person if it's a valid request.
SquadMembership.java is a POJO that holds invited member name and squad ID for
passign back and forth betw FrontController and DAO.

* Implemented join link from 'Squads you are invited to join'

* Implemented ability to decline invitations
Decline link on MySquads now works for squads the user is invited to

* Added create squad functionality
Users can now create a squad using a new Create Squad Link and a createsquad.jsp
Modified FrontController and StepUpDAO accordingly to perform the creation.
Created NewSquadName class to validate proposed squad name input before SQL storage.